ClangFormat: Major formatting change
Current commit consists of major formatting changes to the **entire**
MRPT codebase. The goal of this, is to have MRPT comply with a custom
version of the **Google** style as provided by the ClangFormat tool.

See the docs on instructions on how to sync your work with these changes
prior to making a contribution to the upstream repo

Current commit also aligns the MRPT logo correctly